+++ src/neptune/cli/sync.py
@@ -16,6 +16,7 @@
 
 __all__ = ["SyncRunner"]
 
+import concurrent.futures
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
@@ -50,60 +51,122 @@ logger = get_logger(with_prefix=False)
 
 class SyncRunner:
     @staticmethod
-    def sync_all_offline(*, backend: "NeptuneBackend", base_path: Path, project_name: Optional[str] = None) -> None:
+    def sync_all_offline(
+        *,
+        backend: "NeptuneBackend",
+        base_path: Path,
+        project_name: Optional[str] = None,
+        num_threads: Optional[int] = None,
+    ) -> None:
         containers = collect_containers(path=base_path, backend=backend)
 
-        project = get_project(project_name_flag=QualifiedName(project_name) if project_name else None, backend=backend)
+        project = get_project(
+            project_name_flag=QualifiedName(project_name) if project_name else None,
+            backend=backend,
+        )
         if not project:
             raise CannotSynchronizeOfflineRunsWithoutProject
 
-        for container in containers.offline_containers:
-            container.sync(base_path=base_path, backend=backend, project=project)
+        with concurrent.futures.ThreadPoolExecutor(max_workers=num_threads) as executor:
+            futures = [
+                executor.submit(
+                    container.sync,
+                    base_path=base_path,
+                    backend=backend,
+                    project=project,
+                )
+                for container in containers.offline_containers
+            ]
+            for future in concurrent.futures.as_completed(futures):
+                future.result()
 
     @staticmethod
-    def sync_all(*, backend: "NeptuneBackend", base_path: Path, project_name: Optional[str] = None) -> None:
+    def sync_all(
+        *,
+        backend: "NeptuneBackend",
+        base_path: Path,
+        project_name: Optional[str] = None,
+        num_threads: Optional[int] = None,
+    ) -> None:
         containers = collect_containers(path=base_path, backend=backend)
 
         if containers.unsynced_containers:
-            for async_container in containers.unsynced_containers:
-                async_container.sync(base_path=base_path, backend=backend, project=None)
+            with concurrent.futures.ThreadPoolExecutor(max_workers=num_threads) as executor:
+                futures = [
+                    executor.submit(
+                        async_container.sync,
+                        base_path=base_path,
+                        backend=backend,
+                        project=None,
+                    )
+                    for async_container in containers.unsynced_containers
+                ]
+                for future in concurrent.futures.as_completed(futures):
+                    future.result()
 
         if containers.offline_containers:
             project = get_project(
-                project_name_flag=QualifiedName(project_name) if project_name else None, backend=backend
+                project_name_flag=QualifiedName(project_name) if project_name else None,
+                backend=backend,
             )
             if not project:
                 raise CannotSynchronizeOfflineRunsWithoutProject
 
-            for offline_container in containers.offline_containers:
-                offline_container.sync(base_path=base_path, backend=backend, project=project)
+            with concurrent.futures.ThreadPoolExecutor(max_workers=num_threads) as executor:
+                futures = [
+                    executor.submit(
+                        offline_container.sync,
+                        base_path=base_path,
+                        backend=backend,
+                        project=project,
+                    )
+                    for offline_container in containers.offline_containers
+                ]
+                for future in concurrent.futures.as_completed(futures):
+                    future.result()
 
     @staticmethod
     def sync_selected(
-        *, backend: "NeptuneBackend", base_path: Path, project_name: Optional[str] = None, object_names: Sequence[str]
+        *,
+        backend: "NeptuneBackend",
+        base_path: Path,
+        project_name: Optional[str] = None,
+        object_names: Sequence[str],
+        num_threads: Optional[int] = None,
     ) -> None:
         containers = collect_containers(path=base_path, backend=backend)
-        async_selected = [QualifiedName(name) for name in object_names if not name.startswith(OFFLINE_NAME_PREFIX)]
+        if async_selected := [QualifiedName(name) for name in object_names if not name.startswith(OFFLINE_NAME_PREFIX)]:
+            with concurrent.futures.ThreadPoolExecutor(max_workers=num_threads) as executor:
+                futures = [
+                    executor.submit(
+                        sync_selected_async,
+                        backend=backend,
+                        base_path=base_path,
+                        container_names=[selected],
+                        containers=containers.async_containers,
+                    )
+                    for selected in async_selected
+                ]
+                for future in concurrent.futures.as_completed(futures):
+                    future.result()
 
-        if async_selected:
-            sync_selected_async(
-                backend=backend,
-                base_path=base_path,
-                container_names=async_selected,
-                containers=containers.async_containers,
-            )
-
-        offline_selected = [
+        if offline_selected := [
             UniqueId(name[len(OFFLINE_NAME_PREFIX) :]) for name in object_names if name.startswith(OFFLINE_NAME_PREFIX)
-        ]
-        if offline_selected:
-            sync_selected_offline(
-                backend=backend,
-                base_path=base_path,
-                container_names=offline_selected,
-                containers=containers.offline_containers,
-                project_name=project_name,
-            )
+        ]:
+            with concurrent.futures.ThreadPoolExecutor(max_workers=num_threads) as executor:
+                futures = [
+                    executor.submit(
+                        sync_selected_offline,
+                        backend=backend,
+                        base_path=base_path,
+                        container_names=[selected_offline],
+                        containers=containers.offline_containers,
+                        project_name=project_name,
+                    )
+                    for selected_offline in offline_selected
+                ]
+            for future in concurrent.futures.as_completed(futures):
+                future.result()
 
 
 def sync_selected_async(
@@ -115,11 +178,10 @@ def sync_selected_async(
 ) -> None:
     async_containers_ids = set()
     for container_name in container_names:
-        experiment = get_metadata_container(
+        if experiment := get_metadata_container(
             backend=backend,
             container_id=container_name,
-        )
-        if experiment:
+        ):
             async_containers_ids.add(experiment.id)
         else:
             logger.error(f"Container {container_name} not found")
@@ -138,15 +200,16 @@ def sync_selected_offline(
     containers: List["OfflineContainer"],
     project_name: Optional[str] = None,
 ) -> None:
-    project = get_project(project_name_flag=QualifiedName(project_name) if project_name else None, backend=backend)
+    project = get_project(
+        project_name_flag=QualifiedName(project_name) if project_name else None,
+        backend=backend,
+    )
     if not project:
         raise CannotSynchronizeOfflineRunsWithoutProject
 
     selected_offline_containers: List["OfflineContainer"] = []
     for container_id in container_names:
-        found_container = next((x for x in containers if x.container_id == container_id), None)
-
-        if found_container:
+        if found_container := next((x for x in containers if x.container_id == container_id), None):
             selected_offline_containers.append(found_container)
         else:
             logger.warning("Offline container %s not found on disk.", container_id)